+++ src/cider/nrepl.clj
@@ -272,7 +272,7 @@ Depending on the type of the return value of the evaluation this middleware may 
    "doc-block-tags-fragments"     (str "May be absent. Represent the 'param', 'returns' and 'throws' sections a Java doc comment. " fragments-desc)})
 
 (def info-params
-  {"sym"                "The symbol to lookup"
+  {"sym"                "The symbol to lookup. Must be a string. If it represents a keyword, please express so via the `:symbol-type` param."
    "ns"                 "The current namespace"
    "context"            "A Compliment completion context, just like the ones already passed for the \"complete\" op,
 with the difference that the symbol at point should be entirely replaced by \"__prefix__\".
@@ -280,6 +280,7 @@ For Java interop queries, it helps inferring the precise type of the object the 
 making the results more accurate (and less numerous)."
    "class"              "A Java class. If `:ns` is passed, it will be used for fully-qualifying the class, if necessary."
    "member"             "A Java class member."
+   "symbol-type"        "The type of object of `:sym` as seen by the user. One of: \"symbol\" (default), \"keyword\", \"string\"."
    "var-meta-allowlist" "The metadata keys from vars to be returned. Currently only affects `:clj`.
 Defaults to the value of `orchard.meta/var-meta-allowlist`.
 If specified, the value will be concatenated to that of `orchard.meta/var-meta-allowlist`."})